1. Hibernate orm  2. Bean validation  3. 2 level cache (infinispan)  4. Inhiritance hibernate

+++ src/main/java/com/zhmenko/book/mapper/BookMapper.java
@@ -27,7 +27,7 @@ public abstract class BookMapper {
         return bookEntity.getName();
     }
 
-    public BookEntity bookIdToBookEntity(final Integer bookId) {
+    public BookEntity bookIdToBookEntity(final Long bookId) {
         BookEntity bookEntity = new BookEntity();
         bookEntity.setId(bookId);
         return bookEntity;
