+++ src/main/java/com/zhmenko/exception/AuthorNotFoundException.java
@@ -3,7 +3,7 @@ package com.zhmenko.exception;
 import java.text.MessageFormat;
 
 public class AuthorNotFoundException extends EntityNotFoundException {
-    public AuthorNotFoundException(int id) {
+    public AuthorNotFoundException(Long id) {
         super(MessageFormat.format("Author with id {0} not found!", id));
     }
 }
